# Add the performance test of elementwise_mul