--- a/api/paddle/run.sh
+++ b/api/paddle/run.sh
@@ -3,11 +3,13 @@
 export CUDA_VISIBLE_DEVICES="1"
 #export GLOG_v=4
 
-python abs.py \
+name=${1:-"abs"}
+
+nvprof python ${name}.py \
       --run_with_executor True \
       --check_output False \
       --profiler "none" \
-      --backward False \
+      --backward True \
       --use_gpu True \
-      --repeat 1000 \
+      --repeat 100 \
       --log_level 1