--- a/api/tensorflow/run.sh
+++ b/api/tensorflow/run.sh
@@ -4,10 +4,12 @@ set -xe
 
 export CUDA_VISIBLE_DEVICES="0"
 
-python abs.py \
+name=${1:-"abs"}
+
+nvprof python ${name}.py \
       --check_output False \
       --profile False \
-      --backward False \
+      --backward True \
       --use_gpu True \
       --gpu_id 0 \
       --repeat 100 \